Lab 2 finished 居然不准中文注释;(

+++ lab2/DebugExercise/DebugExercise1.java
@@ -6,7 +6,8 @@ package DebugExercise;
  */
 public class DebugExercise1 {
     public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
+        float quotient = (float) top / bottom;
+        // data type
         int result = Math.round(quotient);
         return result;
     }
